Refactors the existing code.
Move more data related code to the data module and adds tests.

Adds a helper state libarary. This is a hack to work around odd chrashes
observed when using the code directly in a module. Due to resource
manangement bugs in modules quite a number of pointers are leaked. This
is clearly not great, but leaking works around double-free issues. Since
the application is "read-only" this is not a big issue.